refactor: move duplicate and shared functions to models.utils package
