+++ auton_survival/models/cph/__init__.py
@@ -31,8 +31,8 @@ from .dcph_torch import DeepCoxPHTorch, DeepRecurrentCoxPHTorch
 from .dcph_utilities import train_dcph, predict_survival
 
 from auton_survival.utils import _dataframe_to_array
-from auton_survival.models.dsm.dsm_utilities import _get_padded_features
-from auton_survival.models.dsm.dsm_utilities import _get_padded_targets
+from auton_survival.models.utils.recurrent_nn_utils import _get_padded_features
+from auton_survival.models.utils.recurrent_nn_utils import _get_padded_targets
 
 
 class DeepCoxPH:
